Collectd: Rely on the SELinux ansible facts
Currently a first task is run to determine the state of SELinux on the
machine. This task is uncesseary as Ansible provide this information as
part of the facts it gathered.

```
"ansible_selinux": {
    "config_mode": "enforcing",
    "mode": "permissive",
    "policyvers": 28,
    "status": "enabled",
    "type": "targeted"
},
```

Hence the decision to enable or not the collectd_tcp_network_connect
boolean should be based on those facts.

For clarification:

  * config_mode: is the selinux state written in /etc/selinux/config
  * mode: is the selinux state returned by getenforce