RabbitMQ: Allow one to specify interface and port
The playbook as it is right now allow only one to configure RabbitMQ to
listen on all interface and on port 5672.

In a real life deployment when RabbitMQ might be deployed on a server
with more than one NIC, each network having its own purpose,
making it available to all network is not recommended, hence the
operator should have a way to specify on which interface to listen on.

Port is a parameter but wasn't enforced at the configuration level, just
the firewall level. This commit allows on to be able to set the port
value at the configuration level.