Put uchiwa behind Apache
This change configures both uchiwa and kibana to run behind an Apache
frontend, and enforces SSL access to the web UI.  Both applications
can be run on the same host (or on different hosts).  In addition to
changes to the uchiwa, kibana, and httpd roles, this introduces a new
"opstoolsvhost" role that is responsible for a top-level Apache
configuration file.