generate rabbitmq configuration from template
Prior to this commit we were using ansible's `lineinfile` module to
modify /etc/rabbitmq/rabbitmq.config.  This is fragile, in particular
since we were depending on a comment to find our position in the file.
Rather than assuming that the file meets our expectations, this commit
simply replaces it wholesale with a file generated from a template.

Closes #84