[WIP] rewriting overcloud host checks to use systemd
Rewrite the sensu checks that run on the overcloud node to use
(mostly) `systemctl is-active` rather than running oschecks, which
will be moved to the opstools server in a separate commit.

Since most of the services are using a common check, rewrite the
templates so that we can use a single template for the common check
(but allow per-service overrides for things that are not using systemd
or that require some sort of unique configuration).

Change-Id: Ib1fb373aea73fcf9a7ffc8cfeff8f2bc95a91e1a