ensure files are owned by fluentd user
the fluentd package runs the fluentd service as the "fluentd" user,
but our role was installing files with owner "root".  That was
generally not a problem, but when enabling logging over SSL the ssl
private key file is mode 600, which made it inaccessible to fluentd.

This changes the default fluentd_config_owner so that files will be
owned by the fluentd user.

Change-Id: I8fb6024de960e93602f2415fccf57d49bab76bc5