Monitor cinder-volume in case of pacemaker
cinder-volume can be a systemd service or a pacemaker resource. In the
playbook, it was only check as systemd service. With this change, it
check if it's a service or a pacemaker resource.

Warning: in case of pacemaker resource, a sudo rule must be add where
cinder is subscribed as folowing:
/etc/sudoers.d/sensu
---
Defaults:sensu !requiretty

sensu ALL = (root) NOPASSWD:
/usr/libexec/openstack-monitoring/checks/oschecks-pacemaker_host_check
-r *
...

Tested and works over a RHOSP 10 deployment.

Signed-off-by: Cyril Lopez <cylopez@redhat.com>