Prevente to erase user in htpasswd
Basicly if we add a user in htpasswd for uchiwa and kibana manually and
re call ansible playbook it re-create the htpasswd file. So you loose
user create manually. Here I add a check if the file exist and just
update the password of *_proxy_user.

Tested and works on RHEL7.3

Change-Id: I10b4c6f020db3e6dca8b523dd5cbcaa9fcd7c47e
Signed-off-by: Cyril Lopez <cylopez@redhat.com>