gnocchi: use redis when possible
The tooz coordination backend highly recommended is redis. While the
default is the indexer database. For real life workaround using mysql
doesn't work because it consumes too many mysql connection from its
pool.

Also for the incoming driver the redis driver is also recommended.

This change move the coordination and the incoming driver to redis.

Change-Id: Ib597da33081a464a0a4c8c6743af04172bf8b32a